+++ api/models/patient.js
@@ -33,7 +33,16 @@ const PatientSchema = new mongoose.Schema({
     doctor:{
         type:String,
         required :true
+    },
+    password:{
+        type: String,
+        required : true,
     }
+    ,username:{
+        type: String,
+        required : true,
+        unique:true,
+    },
 } )
 
 export default mongoose.model("Patient" , PatientSchema)
