+++ client/src/components/Card.js
@@ -8,12 +8,18 @@ import { Link } from 'react-router-dom';
 
 export default function Cards(props) {
   return (
-    <Card sx={{ background: 'rgba(33, 137, 228, 0.24)', 
+    // <Card sx={{ background: 'rgba(33, 137, 228, 0.24)', 
+    //   borderRadius: '16px',
+    //   boxShadow: '0 4px 30px rgba(0, 0, 0, 0.1)',
+    //   backdropFilter: 'blur(5px)',
+    //   WebkitBackdropFilter: 'blur(5px)',
+    //   border: '1px solid rgba(33, 137, 228, 0.34)',maxWidth:'340px' }}>
+    <Card sx={{ background: props.bg, 
       borderRadius: '16px',
       boxShadow: '0 4px 30px rgba(0, 0, 0, 0.1)',
       backdropFilter: 'blur(5px)',
       WebkitBackdropFilter: 'blur(5px)',
-      border: '1px solid rgba(33, 137, 228, 0.34)',maxWidth:'340px' }}>
+      border: props.border ,maxWidth:'340px' }}>
       <CardActionArea>
         <CardMedia
           component="img"
